Arrange FlatArray From IEnumerable factory test files
